producers/requirements.txt
requests==2.32.4
confluent-kafka==2.11.0
kafka-python==2.0.2